+++ common/src/container/rle/mod.rs
@@ -1,6 +1,9 @@
 //! Tools for working with run length encoded (RLE) data.
 
+use std::{borrow::Borrow, iter::repeat_n};
+
 pub mod bits;
+pub mod downsample;
 pub mod png;
 
 /// Sequence of identical items.
@@ -10,12 +13,40 @@ pub struct Run<T = u8> {
     pub value: T,
 }
 
+impl<T> Run<T> {
+    pub fn new(length: u64, value: T) -> Self {
+        Self { length, value }
+    }
+}
+
 /// Decode a RLE sequence into a mutable slice.
-pub fn decode_into<T: Clone>(decoder: impl Iterator<Item = Run<T>>, image: &mut [T]) {
+pub fn decode_into<T, R, D>(decoder: D, image: &mut [T])
+where
+    T: Clone,
+    R: Borrow<Run<T>>,
+    D: IntoIterator<Item = R>,
+{
     let mut pixel = 0;
     for run in decoder {
+        let run = run.borrow();
         let length = run.length as usize;
-        image[pixel..(pixel + length)].fill(run.value);
+        image[pixel..(pixel + length)].fill(run.value.clone());
         pixel += length;
     }
+}
+
+pub fn decode_vec<T, R, D>(decoder: D) -> Vec<T>
+where
+    T: Clone,
+    R: Borrow<Run<T>>,
+    D: IntoIterator<Item = R>,
+{
+    let mut out = Vec::new();
+    for run in decoder {
+        let run = run.borrow();
+        let length = run.length as usize;
+        out.extend(repeat_n(run.value.clone(), length));
+    }
+
+    out
 }