+++ slicer/src/slicer/slice_raster.rs
@@ -1,19 +1,20 @@
 // Uses an Active Edge Table (AET) for faster filling.
 // Reference: https://www.cs.rit.edu/~icss571/filling/how_to.html
 
-use std::{
-    collections::VecDeque,
-    sync::atomic::{AtomicU64, Ordering},
-};
+use std::collections::VecDeque;
 
 use common::{
-    slice::{EncodableLayer, SliceResult},
+    container::{
+        Run,
+        rle::downsample::{downsample, downsample_adjacent},
+    },
+    slice::Layer,
     units::Milimeter,
 };
 use itertools::Itertools;
-use nalgebra::{Vector2, Vector3};
+use nalgebra::Vector2;
 use ordered_float::OrderedFloat;
-use rayon::iter::{IntoParallelIterator, ParallelIterator};
+use rayon::iter::{IndexedParallelIterator, IntoParallelIterator, ParallelIterator};
 
 use crate::{
     geometry::Segments1D,
@@ -22,10 +23,11 @@ use crate::{
 
 impl Slicer {
     /// Actually runs the slicing operation, it is multithreaded.
-    pub fn slice_raster<Layer: EncodableLayer>(&self) -> SliceResult<'_, Layer::Output> {
-        let platform = self.slice_config.platform_resolution;
-        let pixels = (platform.x * platform.y) as u64;
-        let voxels = AtomicU64::new(0);
+    pub fn slice_raster(&self) -> Vec<Layer> {
+        let supersample = self.slice_config.supersample;
+        let real_platform = self.slice_config.platform_resolution;
+        let platform = real_platform * supersample as u32;
+        let pixels = platform.x as u64 * platform.y as u64;
 
         // A segment contains a reference to all of the triangles it contains. By
         // splitting the mesh into segments, not all triangles need to be tested
@@ -35,11 +37,11 @@ impl Slicer {
             .map(|model| Segments1D::from_mesh(&model.mesh, SEGMENT_LAYERS))
             .collect::<Vec<_>>();
 
-        let layers = (0..self.layers)
+        (0..self.layers * supersample as u32)
             .into_par_iter()
             .map(|layer| {
-                let height = layer as f32 * self.slice_config.slice_height.get::<Milimeter>();
-                let mut voxels_inner = 0;
+                let height = layer as f32 / supersample as f32
+                    * self.slice_config.slice_height.get::<Milimeter>();
 
                 // Gets all the intersections between the slice plane and the
                 // model. Because all the faces are triangles, every triangle
@@ -47,30 +49,31 @@ impl Slicer {
                 // interpreted as line segments making up a polygon.
                 let segments = self.models.iter().enumerate().flat_map(|(idx, model)| {
                     let intersections = segments[idx].intersect_plane(&model.mesh, height);
-                    (intersections.into_iter()).map(|segment| (segment, model.exposure))
+                    intersections.into_iter().map(|(pos, dir)| {
+                        ((pos.map(|x| x * supersample as f32), dir), model.exposure)
+                    })
                 });
+
                 let mut edges = global_edge_table(segments);
-
-                // Creates a new encoded for this layer. Because printers can
-                // have very high resolution displays, the uncompressed data for
-                // a sliced model can easily be over 30 Gigabytes. Most formats
-                // use some sort of compression scheme to resolve this issue, so
-                // to use a little memory as needed, the layers are compressed
-                // as they are made.
-                let mut encoder = Layer::new(platform);
-                let mut last = 0;
-
                 let mut active = Vec::new();
-                let mut y = edges.front().map(|e| e.min.y).unwrap_or(0);
+                let first_y = edges.front().map(|e| e.min.y).unwrap_or(0);
 
+                let mut runs = Vec::new();
+                let padding = (first_y as u64 / supersample as u64) * real_platform.x as u64;
+                runs.push(Run::new(padding, 0));
+
+                let mut rows = vec![Vec::new(); supersample as usize];
+                let mut row = Vec::new();
+                let mut y = first_y;
                 while (!edges.is_empty() || !active.is_empty()) && y < platform.y {
                     update_active_edges(&mut edges, &mut active, y);
-                    let y_offset = (platform.x * y) as u64;
+                    let out = if supersample > 1 { &mut row } else { &mut runs };
 
                     // Convert the intersections into runs of voxels to be
                     // encoded into the layer.
                     let mut depth = 0;
                     let mut exposures = [0; 256];
+                    let mut last = 0;
                     for (a, b) in active.iter().tuple_windows() {
                         let delta = 1 - (a.direction as i32) * 2;
                         depth += delta;
@@ -78,14 +81,27 @@ impl Slicer {
 
                         let [a, b] = [a.x, b.x].map(|x| (x.round() as u64).min(platform.x as u64));
                         if depth != 0 && b != a {
-                            let (start, length) = (a + y_offset, b - a);
-                            (start > last).then(|| encoder.add_run(start - last, 0));
+                            let (start, length) = (a, b - a);
+                            (start > last).then(|| out.push(Run::new(start - last, 0)));
 
                             let exposure = exposures.iter().rposition(|&x| x > 0).unwrap_or(255);
-                            encoder.add_run(length, exposure as u8);
-
-                            voxels_inner += length;
+                            out.push(Run::new(length, exposure as u8));
                             last = start + length;
+                        }
+                    }
+
+                    // Fill the empty space at the end of the row
+                    let padding = platform.x as u64 - last;
+                    (padding > 0).then(|| out.push(Run::new(padding, 0)));
+
+                    if supersample > 1 {
+                        let ss_row = ((y - first_y) % supersample as u32) as usize;
+                        downsample_adjacent(supersample, &row, &mut rows[ss_row]);
+                        row.clear();
+
+                        if (y + 1) % supersample as u32 == 0 {
+                            downsample(&rows, platform.x as u64, &mut runs);
+                            rows.iter_mut().for_each(Vec::clear);
                         }
                     }
 
@@ -96,21 +112,27 @@ impl Slicer {
                 // decoded into is just uninitialized memory. So if the last run
                 // doesn't fill the buffer, the printer will just print whatever
                 // was in the buffer before which just makes a huge mess.
-                (last < pixels).then(|| encoder.add_run(pixels - last, 0));
+                if y < platform.y {
+                    let rows = (real_platform.y - y / supersample as u32) as u64;
+                    runs.push(Run::new(rows * real_platform.x as u64, 0));
+                }
 
-                // Finished encoding the layer
-                voxels.fetch_add(voxels_inner, Ordering::Relaxed);
-                encoder.finish(layer, &self.slice_config)
+                runs
+            })
+            .chunks(supersample as usize)
+            .enumerate()
+            .map(|(i, mut chunk)| {
+                let data = if supersample > 1 {
+                    downsample_to_vec(&chunk, pixels)
+                } else {
+                    chunk.pop().unwrap()
+                };
+
+                let exposure = self.slice_config.exposure_config(i as u32).clone();
+                Layer { data, exposure }
             })
             .inspect(|_| self.progress.add_complete(1))
-            .collect::<Vec<_>>();
-
-        self.progress.set_finished();
-        SliceResult {
-            layers,
-            voxels: voxels.load(Ordering::Relaxed),
-            slice_config: &self.slice_config,
-        }
+            .collect::<Vec<_>>()
     }
 }
 
@@ -135,7 +157,7 @@ struct ActiveEdge {
 }
 
 fn global_edge_table(
-    segments: impl Iterator<Item = (([Vector3<f32>; 2], bool), u8)>,
+    segments: impl Iterator<Item = (([Vector2<f32>; 2], bool), u8)>,
 ) -> VecDeque<Edge> {
     let mut edges = Vec::new();
     for ((pos, direction), exposure) in segments {
@@ -148,7 +170,7 @@ fn global_edge_table(
         }
 
         edges.push(Edge {
-            min: pos[(pos[0].y >= pos[1].y) as usize].xy(),
+            min: pos[(pos[0].y >= pos[1].y) as usize],
             y_max: pos[0].y.max(pos[1].y),
             inv_slope: dx / dy,
             direction,
@@ -174,4 +196,10 @@ fn update_active_edges(edges: &mut VecDeque<Edge>, active: &mut Vec<ActiveEdge>,
         });
     }
     active.sort_by_key(|x| OrderedFloat(x.x));
+}
+
+fn downsample_to_vec(chunks: &[Vec<Run>], width: u64) -> Vec<Run> {
+    let mut out = Vec::new();
+    downsample(chunks, width, &mut out);
+    out
 }