+++ slicer/src/util.rs
@@ -1,50 +1,51 @@
-use common::{
-    container::Image,
-    slice::{DynSlicedFile, EncodableLayer, Format, SliceConfig, SliceResult},
-};
+use std::borrow::Borrow;
 
-pub fn export(config: &SliceConfig, layers: impl Iterator<Item = Image>) -> (DynSlicedFile, u64) {
+use common::slice::{self, DynSlicedFile, EncodableLayer, Format, SliceConfig};
+
+pub fn export_raster<Layers, Layer>(
+    config: &SliceConfig,
+    layers: Layers,
+    voxels: u64,
+) -> DynSlicedFile
+where
+    Layers: IntoIterator<Item = Layer>,
+    Layer: Borrow<slice::Layer>,
+{
     match config.format {
-        Format::Goo => {
-            let result = encode_layers::<goo_format::LayerEncoder>(config, layers);
-            let voxels = result.voxels;
-            let file = Box::new(goo_format::File::from_slice_result(result));
-            (file, voxels)
-        }
-        Format::Ctb => {
-            let result = encode_layers::<ctb_format::LayerEncoder>(config, layers);
-            let voxels = result.voxels;
-            let file = Box::new(ctb_format::File::from_slice_result(result));
-            (file, voxels)
-        }
-        Format::NanoDLP => {
-            let result = encode_layers::<nanodlp_format::LayerEncoder>(config, layers);
-            let voxels = result.voxels;
-            let file = Box::new(nanodlp_format::File::from_slice_result(result));
-            (file, voxels)
-        }
+        Format::Goo => Box::new(goo_format::File::from_layers(
+            config,
+            encode_raster_layers::<goo_format::LayerEncoder, _, _>(config, layers),
+        )),
+        Format::Ctb => Box::new(ctb_format::File::from_layers(
+            config,
+            encode_raster_layers::<ctb_format::LayerEncoder, _, _>(config, layers),
+        )),
+        Format::NanoDLP => Box::new(nanodlp_format::File::from_layers(
+            config,
+            encode_raster_layers::<nanodlp_format::LayerEncoder, _, _>(config, layers),
+            voxels,
+        )),
         Format::Svg => panic!(),
     }
 }
 
-fn encode_layers<Layer: EncodableLayer>(
-    slice_config: &SliceConfig,
-    layers: impl Iterator<Item = Image>,
-) -> SliceResult<'_, Layer::Output> {
-    let mut voxels = 0;
-    let layers = layers.into_iter().enumerate().map(|(i, image)| {
-        let mut encoder = Layer::new(slice_config.platform_resolution);
-        for run in image.runs() {
-            encoder.add_run(run.length, run.value);
-            (run.value > 0).then(|| voxels += run.length);
-        }
-
-        encoder.finish(i as u32, slice_config)
-    });
-
-    SliceResult {
-        layers: layers.collect(),
-        slice_config,
-        voxels,
-    }
+pub fn encode_raster_layers<Encoder, Layers, Layer>(
+    config: &SliceConfig,
+    layers: Layers,
+) -> Vec<Encoder::Output>
+where
+    Encoder: EncodableLayer,
+    Layers: IntoIterator<Item = Layer>,
+    Layer: Borrow<slice::Layer>,
+{
+    layers
+        .into_iter()
+        .enumerate()
+        .map(|(i, layer)| {
+            let layer = layer.borrow();
+            let mut encoder = Encoder::new(config.platform_resolution);
+            (layer.data.iter()).for_each(|run| encoder.add_run(run.length, run.value));
+            encoder.finish(i as u32, config, &layer.exposure)
+        })
+        .collect()
 }